calculation_service/requirements.txt
django
djangorestframework
pyyaml
uritemplate
coreapi